+++ NEWS.md
@@ -1,7 +1,7 @@
 ## RStudio 2026.10.0 "Blue Mistflower" Release Notes
 
 ### New
--
+- Optimized RStudio Desktop startup time.
 
 ### Fixed
 -